Cleanup and regroup commands, new tabline theme
The buffer-/tabline is very confusing under the the original monokai
theme used by airline when it comes to discriminating the current,
alternative and modified buffers. Therefore the theme was switched.

A lot of tab related commands that were part of the original clone
that this vimrc is based on where totally unused, because tabs are
a feature that is very rarely used by myself, so in the interest of
a lean vimrc they were scrapped.

Additionally some commands where put into different sections in order
to arrive at a more consistent grouping.