Add active window highlighting
When using multiple windows (i.e. using splits) it can sometimes get
confusing which window is the currently active one. Usually the
statusline is an identifier, depending on the used colorscheme it
nevertheless can be not immediately apparent.

Therefore the cursorline, highlighting the current line was restricted
to be shown in the active window only. Similarly, relative linenumbers
are only shown in the active window as there is no point in using this
feature anywhere else. This was achieved by autocommands hooking into
the appropriate events.