Intuitive ex-cmd-completion and add plugin fzf
The previously used 'wildmode=list' option for ex-mode command
completion had a tendenecy to stack on itself, leading to a vertically
evergrowing command line, which was confusing at times. The simpler
'wildmenu' used instead has the additional benefit that the currently
completed selection is highlighted in the list.

FZF is a useful fuzzy completer for filenames/directories/tags/...
Usually a commandline tool, it also integrates nicely with vim.