vimrc: Add commands for buffers, path of current file and typo-helpers
buffer related:
Two abbreviations added to allow for faster switching and deleting of
buffers by displaying the current bufferlist when typing `:b<Space>` or
`bd<Space>`.

file-path related:
When working on a project, vim's working directory commonly sits at the
root of the project tree. In such situations it can be quite annoying to
open multiple files in the same subdirectory, as the relative path to
this subdirectory has to be specified each time `:edit` is invoked. For
this kind of situation the new `%%` mapping has been introducedm, which
expands to the path of the current file. Try `:edit %%<TAB>` to see this
in action.

common typos:
A lot of times when closing a window or writing the buffer to disk, it
happens that the shift key used to type the colon doesn't get released
before the subsequent character is typed, resulting in commands of the
form `:Q*` or `:W*`. Thus a couple of custom commands mapping the most
common of these kinds of typos to their default uncapitalized variants
have been introduced.