Add conda initialization to bashrc
The package manager conda, part of the python distribution Anaconda, needs to
perform some path manipulation in order to setup its environments. The init
function that achieves this, has been added to the bashrc automatically through
the install script of Anaconda and can therefore be removed should it ever be
uninstalled.