bashrc: disable terminal flow control when interactive
The XON/XOFF signals sent when pressing <C-q> / <C-s>, causing to
continue/stop the terminal have been disabled as they tend to be a
nuisance when pressed accidentally.

Additionally the option for colored gcc warnings (found in the default
bashrc of Ubuntu 18.04+) has been included, though it is still left
commented.