vimrc: disable some custom buffer commands
The custom functions BClose and BDelete and their mappings were
shadowing the <Leader>b mapping for fzf's :Buffers, causing it to be
triggered only after a timeout. To get rid of this irritating delay the
seldomly used BClose and BDelete were unmapped, they can likely be
removed entirely in a future cleanup of this file.