makesymlinks.sh: dynamic handling of repository contents
Previously all files to be linked had to be specified in the source text of
`makesymlinks.sh`. The setup script has been rewritten such that it should glob
the entire contents of the repository and create symlinks in the appropriate
postions in $HOME.

Additionally files already present in $HOME will interactively prompt before
being removed by the call to `ln`.