vimrc: Add LocalLeader <C-Space>
This removes all old mappings of <C-@>, in particular the ones assigned to
latex-suite. Therefore the mapppings for latex-suite's jump command, together
with the movementes for changing windows, have been restored back to their
defaults. This in turn freed the command <C-L> again, which has now been
enhanced in order to not only redraw the screen but also clear the search
highlight.