vimrc: replace vim-man with default man plugin
A casual browse of the vim help pages revealed that vim ships with
support for reading man pages by default, the plugin just needs to be
loaded from the `ftplugin` folder of the installation.

The plugin has been configured to open vertical splits by default and
act as the standard `keywordprg`, i.e. the program used for lookup when
pressing `K`.