vimrc: Disable macro recording in cmdline-windows
The cmdline window (entered through `q:`, `q/` or `q?`) is sometimes
opened accidentally, particularly when mistyping `:q` for quitting. The
situation usually worsens when a second typo then starts the recording
of a macro through `q`. To prevent this confusing behaviour and since
macro recording inside a cmdline window does not seem all that useful,
the ability to record macros has been disabled inside of cmdline window
via autocommand mappings. Instead, pressing `q` inside a cmdline window
will now close it. If this behaviour turns out to be desirable, it can
be used for other special buffers, like the quickfix-list.