vimrc: Map <C-]> to use :tjump
When multiple tags are present, <C-]> originally just used to jump to
the first tag and did not report on the other tags being present. With
this remap, the behavior is changes such that in the presenece of
multiple tags the tselect menu will be brought up.

Additionally the mapping involving <Cmd> had to be replaced by a colon,
as apparently <Cmd> is a recent addition, so it is not understood by
older versions of vim.