vim: Add Zotero integration and use biber backend
A running Zotero instance can be queried for cite keys to be added in
the open buffer. This functionality is bound to `<Leader>z` for now.

The ftplugins for latex files have been modified to use the biber
backend for the reference generation as well as better handling of the
various generated files by recognizing the suffixes.