bash: Add standard completion path for local applications
Applications installed to `~/.local/bin` sometimes ship with bash
completion scripts. These can now be added/symlinked to
`~/.bash_completion.d` and will be picked up and sourced by the shell.

This is also useful for aliases, which usually get ignored by the
completion functionality and have to be reregistered with `complete`.
Therefore this addition was placed at the top of the `.bash_aliases`
file, so that we can make sure that all necessary completion functions
have already been loaded.