vimrc: Manage fzf entirely through vim-plug
The previously registerd post-update hook for fzf was using the install
script implicitly as `./install --bin`, which causes only a download of
the fzf binary, but no update of `$PATH` or registering of
bash-completions. This was changed back to the 'old' (non cross-platform
compatible, i.e. won't work on MS-windows) hook, which calls the install
script with the `--all` argument instead.