Update makesymlinks.sh to work off of HEAD
Previously the stow script got its list of packages to stow by looking
at the `master` branch. This can potentially lead to problems once
additional branches are being used or the `HEAD` is detached to an older
commit.

Therefore the list of packages to be stowed is now derived from `HEAD`
instad of `master`.