vimrc: Fix toggle checklist remap issue of vimwiki
The toggle checklist functionality of vimwiki couldn't be remapped by a
custom mapping due to a regression in a commit contained in the `master`
branch. After digging into the root cause, it was realized that the
issue is already resolved in the `dev` branch of the repository,
therefore the plugin was switched to the dev branch.