vim: Remove special hlsearch case for tagbar
Previously the highlighted search was felt to be annoying in the tagbar
and I thought I could remove it there via an ftplugin that containing a
`setlocal` command for this option. Turns out that `hlsearch` is global,
so the highlight for the searched term was disabled for the entire vim
whenever entering the tagbar. Therefore this option was commented out
for now.