vimrc: Fix minor bug with active window highlight
Forcefully set the relativenumber, cursorline and colorcolumn on
startup, as it can otherwise cause the active window to be missing
these. The way the autoevent handling and function are written, there
are likely still some edge cases that need to be ironed out, such as
when interacting with popup windpws (fzf, ...) and terminal windows.